requirements.txt
Django==1.10.4
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
django-cors-headers==1.3.1
selenium==3.0.2
Pillow==4.0.0
